+++ src/bittytax/conv/parsers/binance.py
@@ -362,7 +362,10 @@ def _parse_binance_statements_row(
             buy_asset=row_dict["Coin"],
             wallet=WALLET,
         )
-    elif row_dict["Operation"] == "Distribution":
+    elif row_dict["Operation"] in (
+        "Distribution",
+        "Token Swap - Distribution",
+    ):
         if Decimal(row_dict["Change"]) > 0:
             data_row.t_record = TransactionOutRecord(
                 TrType.AIRDROP,